requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.1.0
attrs==23.1.0
blinker==1.7.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.10.0
deprecation==2.1.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
fonttools==4.46.0
gotrue==1.3.1
greenlet==3.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.1.4
Pillow==9.5.0
pluggy==1.3.0
postgrest==0.10.8
psutil==5.9.6
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==2.4.7
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
realtime==1.0.2
referencing==0.32.0
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.7
rpds-py==0.13.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
supervision==0.17.1
sympy==1.12
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.134
urllib3==2.1.0
validators==0.22.0
websockets==11.0.3
Werkzeug==2.3.7
